+++ src/Resources/Policy.Autorest/Properties/AssemblyInfo.cs
@@ -33,3 +33,12 @@ using System.Runtime.InteropServices;
 
 
 
+
+
+
+
+
+
+
+
+