+++ src/Resources/Policy.Autorest/test/Common.ps1
@@ -98,6 +98,7 @@ if ($TargetTestName -and (!$TestName -or ($TestName -eq $TargetTestName))) {
         $policyAssignmentMissingLocation = $env.policyAssignmentMissingLocation
         $policyAssignmentMissingIdentityId = $env.policyAssignmentMissingIdentityId
         $policyExemptionNotFound = $env.policyExemptionNotFound
+        $policyEnrollmentNotFound = $env.policyEnrollmentNotFound
         $invalidRequestContent = $env.invalidRequestContent
         $policyDefinitionParameter = $env.policyDefinitionParameter
         $missingSubscription = $env.missingSubscription