+++ src/Resources/Resources/ChangeLog.md
@@ -23,7 +23,9 @@
     - Fixed dynamic parameter inheritance in deployment and deployment stack cmdlets
 * Updated deployment stack cmdlets to `2025-07-01` API; added `-ResourcesWithoutDeleteSupport`, `-ValidationLevel`
 * Fixed `Set-AzRoleAssignment` unable to delete conditions
-* Updated Policy.Autorest to 2025-03-01: added `-Expand`, `-EnforcementMode`, `-Version`; removed `-BackwardCompatible`
+* Updated Policy.Autorest to 2026-01-01
+    - Added `-Expand`, `-EnforcementMode`, `-Version`; removed `-BackwardCompatible`
+    - Added `Get-AzPolicyEnrollment`, `New-AzPolicyEnrollment`, `Remove-AzPolicyEnrollment`, and `Update-AzPolicyEnrollment` cmdlets to interact with new Policy Enrollments resources.
 * Fixed `Get-AzRoleDefinition` null ABAC `Condition` [#29058] [#25940]
 * [Breaking Change] Role definition cmdlets use `Permissions` array with per-permission conditions